+++ example/lib/main.dart
@@ -24,54 +24,44 @@ class App extends StatelessWidget {
             padding: const EdgeInsets.all(32.0),
             child: Column(
               crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                const Text('Glitters'),
-                const SizedBox(height: 4.0),
+              children: const [
+                Text('Glitters'),
+                SizedBox(height: 4.0),
                 Expanded(
                   flex: 2,
-                  child: ColoredBox(
-                    color: Colors.blueGrey.shade700,
-                    child: Stack(
-                      children: const <Widget>[
-                        Glitters(
-                          minSize: 8.0,
-                          maxSize: 20.0,
-                          interval: Duration.zero,
-                          maxOpacity: 0.7,
-                        ),
-                        Glitters(
-                          minSize: 10.0,
-                          maxSize: 25.0,
-                          interval: Duration(milliseconds: 20),
-                          color: Colors.lime,
-                        ),
-                        Glitters(
-                          minSize: 10.0,
-                          maxSize: 25.0,
-                          duration: Duration(milliseconds: 200),
-                          inDuration: Duration(milliseconds: 500),
-                          outDuration: Duration(milliseconds: 500),
-                          interval: Duration(milliseconds: 30),
-                          color: Colors.white,
-                          maxOpacity: 0.7,
-                        ),
-                        Glitters(
-                          minSize: 14.0,
-                          maxSize: 30.0,
-                          interval: Duration(milliseconds: 40),
-                          color: Colors.orange,
-                        ),
-                      ],
-                    ),
+                  child: GlitterStack(
+                    backgroundColor: Color(0xFF455A64),
+                    minSize: 15.0,
+                    maxSize: 30.0,
+                    interval: Duration(milliseconds: 300),
+                    duration: Duration(milliseconds: 300),
+                    inDuration: Duration(milliseconds: 300),
+                    outDuration: Duration(milliseconds: 300),
+                    // maxOpacity: 0.5,
+                    children: [
+                      Glitters(),
+                      Glitters(
+                        color: Colors.lime,
+                        delay: Duration(milliseconds: 300),
+                      ),
+                      Glitters(
+                        color: Colors.white,
+                        delay: Duration(milliseconds: 600),
+                      ),
+                      Glitters(
+                        color: Colors.orange,
+                        delay: Duration(milliseconds: 900),
+                      ),
+                    ],
                   ),
                 ),
-                const SizedBox(height: 36.0),
-                const Text('SingleGlitter'),
-                const SizedBox(height: 4.0),
+                SizedBox(height: 36.0),
+                Text('SingleGlitter'),
+                SizedBox(height: 4.0),
                 Expanded(
                   child: ColoredBox(
-                    color: Colors.blueGrey.shade700,
-                    child: const SingleGlitter(
+                    color: Color(0xFF455A64),
+                    child: SingleGlitter(
                       color: Colors.amber,
                     ),
                   ),